refactor(view): P6-A3 first cut — extract design_tokens.cpp from design_import.cpp
Phase 6 A3 (design_import.cpp split), first cut. The design-token
import/export surface — the cluster that converts between Pulp's
Theme and three external token formats — relocated into its own TU
`core/view/src/design_tokens.cpp` (558 lines).

Moved (definitions only; declarations stay in design_import.hpp):
  * W3C Design Tokens     — parse_w3c_tokens / export_w3c_tokens
  * Figma Variables       — parse_figma_variables / export_figma_variables
  * Stitch Design System  — parse_stitch_design_system / export_stitch_design_system
  * Theme ⇄ IRTokens      — ir_tokens_to_theme / theme_to_ir_tokens
  * file-local parse_hex_color_str / parse_design_number helpers

The token parsers reference design_import.cpp's file-local JSON
accessors (get_string / get_float) — duplicated into design_tokens.cpp's
anonymous namespace so the new TU stands alone. The now-unused get_int
helper is removed from design_import.cpp.

design_import.cpp: 4,691 → 4,155 (-536). Token-format work no longer
recompiles the whole 4.7k-line importer. Subsequent A3 cuts will peel
the code-generator and claude-bundle clusters.

Verified: pulp-view builds clean; pulp-test-design-import (30 cases)
+ pulp-test-design-import-w3c-tokens (12 cases) all pass.
